application/pre_deploy/requirements.txt
opensearch-py==2.4.2
numpy
pandas
python-dotenv
tqdm
openai
weaviate-client